requirements.txt
asgiref==3.5.2
django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
matplotlib==3.5.1
numpy==1.21.5
pandas==1.4.1
Pillow==9.1.0
PyJWT==2.5.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
